Liberated the tape parser template.
